rename file and fix syntax
