+++ src/aiod/__init__.py
@@ -1,5 +1,5 @@
 from importlib.metadata import version
-from aiod.default.authorization_test import get_current_user
+from aiod.default.get_current_user import get_current_user
 from aiod.resources import case_studies
 from aiod.resources import computational_assets
 from aiod.resources import contacts
